# perf: make loading shallow weights up to 20X faster