# What does this PR do?

<!--
Congratulations! You've made it this far! You're not quite done yet though.

Once merged, your PR is going to appear in the release notes with the title you set, so make sure it's a great title that fully reflects the extent of your awesome contribution.

Then, please replace this with a description of the change and which issue is fixed (if applicable). Please also include relevant motivation and context. List any dependencies (if any) that are required for this change.

Once you're done, someone will review your PR shortly (see the section "Who can review?" below to tag some potential reviewers). They may suggest changes to make the code even better. If no one reviewed your PR after a week has passed, don't hesitate to post a new comment @-mentioning the same persons---sometimes notifications get lost.
-->

<!-- Remove if not applicable -->

Fixes #361. Local profiling shows majority of the time is spent on recurring accesses to the tensor metadata on every individual tensor fetching. This might be negligible when fetching a few tensors (and is actually better to have it lazily, so I understand the design intentions) but when reading whole weight files (such as a diffusion LoRA) it adds up quickly and causes exponential performance degradation.

With this PR, the example in #361 is now as fast as torch (and performs even better)! Median went from 0.33 seconds to 0.017 seconds. Torch is still at 0.023 seconds so safetensors beats it :) 

Testing a [real world LoRA](https://huggingface.co/nerijs/pixel-art-xl/blob/main/pixel-art-xl.safetensors) with results in similiar perf gains ([script.py](https://gist.github.com/isidentical/0739bfa3d0c5cb0d6946a7561d7c8171)) where median goes from ~0.4 seconds to under 0.018 seconds (more than 20X).