@isidentical 2023-09-16
Benchmarks:
```
---------------------------------------------------------------------------------------------------------------------------- benchmark: 4 tests ---------------------------------------------------------------------------------------------------------------------------
Name (time in ms)                                                                                   Min                   Max                  Mean             StdDev                Median                IQR            Outliers       OPS            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_pt_sf_load_cpu[weight_creator0] (NOW)                                                       1.3438 (1.0)         33.3606 (12.41)        1.5071 (1.0)       1.9692 (111.51)       1.3747 (1.0)       0.0184 (1.0)          2;43  663.5463 (1.0)         511           1
test_pt_sf_load_cpu[weight_creator0] (.benchmarks/Linux-CPython-3.11-64bit/0003_baselin)         2.5957 (1.93)         2.6880 (1.0)          2.6348 (1.75)      0.0177 (1.0)          2.6320 (1.91)      0.0223 (1.21)        103;5  379.5325 (0.57)        329           1
test_pt_sf_load_cpu[weight_creator1] (NOW)                                                      34.9504 (26.01)       73.6812 (27.41)       55.2946 (36.69)    19.0215 (>1000.0)     70.4852 (51.27)    37.2772 (>1000.0)       4;0   18.0850 (0.03)         13           1
test_pt_sf_load_cpu[weight_creator1] (.benchmarks/Linux-CPython-3.11-64bit/0003_baselin)     1,493.7300 (>1000.0)  1,534.2338 (570.77)   1,505.5897 (999.03)   16.4876 (933.67)   1,499.5223 (>1000.0)  16.1243 (877.93)        1;0    0.6642 (0.00)          5           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```