@Narsil 2023-09-18
Superseed by #363 

Thanks @isidentical for confirming the other PR works just as good and creating this !